requirements.txt
google-cloud-speech
vertexai
simpleaudio
gtts
pydub
